refactor: consolidate classes instead of inheritance
